+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [6.0.0] - 2020-07-08
+
+### Added
+
+- Manipulation: RegionGenerator, CartesianWellGenerator
+
+### Changed
+
+- Using boost::unordered_map
+- WellGenerator well start point must be given in wkt format
+
+[6.0.0]: https://github.com/felipegiacomelli/MSHtoCGNS/compare/v5.0.0...v6.0.0
+
+## [5.0.0] - 2020-06-06
+
+### Changed
+
+- Rolled back to CGNS-3.4.0 due to incompatibility issues between v4.1.1 and v3.4.0
+
+[5.0.0]: https://github.com/felipegiacomelli/MSHtoCGNS/compare/v4.0.0...v5.0.0
+
 ## [4.0.0] - 2020-04-18
 
 ### Added